Remove partial retrieval of pocket articles
Previously the PocketWorker was only pulling the articles that were
added or modified since the last time it checked. This lowered the
bandwidth cost of each poll, but added a lot of complexity to the
process of merging the newly retrieved results with the results that the
Worker already had in redis. By pulling all archived articles each time,
we can avoid the complexity of merging the two lists, at the cost of a
larger download on each poll, which is a price worth paying for now.