Ignore error in `now rm`
The rm command allows us to cleanup after a new deploy, but if deploying
is a no-op, rm will exit with an error since it didn't remove anything.

This ignores any error from rm. Not great since it's a catch-all but
it's the only thing I can think of right now.